Renamed duplicated file name.
