+++ tio/CMakeLists.txt
@@ -13,5 +13,5 @@ TARGET_INCLUDE_DIRECTORIES(tio PUBLIC ${PROJECT_SOURCE_DIR}/include)
 SET(version "1.0.0")
 SET(soversion "1.0.0")
 SET_TARGET_PROPERTIES(tio PROPERTIES VERSION ${version} SOVERSION ${soversion})
-INSTALL(FILES include/tio.h include/tio2.h ../khc/include/khc.h ../khc/include/khc_socket_callback.h ../kii/kii.h ../kii/kii_json_utils.h ../kii/kii_task_callback.h ../kii_json/include/kii_json.h ../kii_json/libs/jsmn/jsmn.h DESTINATION include)
+INSTALL(FILES include/tio.h include/tio2.h ../khc/include/khc.h ../khc/include/khc_socket_callback.h ../kii/kii.h ../kii/kii_task_callback.h ../kii_json/include/kii_json.h ../kii_json/libs/jsmn/jsmn.h DESTINATION include)
 INSTALL(TARGETS tio LIBRARY DESTINATION lib)
